# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AiMedRes Medical AI API
  description: |
    API for AI-driven medical research and clinical decision support.
    
    **Security:** All endpoints require authentication via API key or OAuth2 token.
    **Compliance:** HIPAA/GDPR compliant with PHI de-identification and audit logging.
  version: 1.0.0
  contact:
    name: AiMedRes Support
    url: https://github.com/V1B3hR/AiMedRes
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://staging.aimedres.example.com/api/v1
    description: Staging server
  - url: https://api.aimedres.example.com/api/v1
    description: Production server
security:
  - ApiKeyAuth: []
  - OAuth2: [openid, profile, email]
...
paths:
...
  /explain/attribution:
    post:
      summary: Feature attribution
      description: Get feature attributions for model prediction
      operationId: featureAttribution
      tags: [Explainability]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExplainRequest'
      responses:
        '200':
          description: Attribution computed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttributionResponse'
  /explain/uncertainty:
    post:
      summary: Uncertainty estimation
      description: Get uncertainty and confidence metrics for prediction
      operationId: uncertaintyEstimation
      tags: [Explainability]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExplainRequest'
      responses:
        '200':
          description: Uncertainty computed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UncertaintyResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: /auth/oidc/authorize
          tokenUrl: /auth/oidc/token
          scopes:
            openid: OpenID Connect
            profile: User profile
            email: User email
  schemas:
...
    ExplainRequest:
      type: object
      required: [prediction_id]
      properties:
        prediction_id:
          type: string
        case_id:
          type: string
    AttributionResponse:
      type: object
      properties:
        attributions:
          type: array
          items:
            type: object
            properties:
              feature:
                type: string
              importance:
                type: number
              value:
                type: number
    UncertaintyResponse:
      type: object
      properties:
        confidence:
          type: number
        uncertainty:
          type: number
        method:
          type: string
...